implementazione test della rimozione con exceptio e della ral superiore a 30000

+++ src/main/java/it/prova/gestionesocieta/service/SocietaServiceImpl.java
@@ -4,12 +4,14 @@ import java.util.List;
 
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
+import javax.persistence.TypedQuery;
 
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import it.prova.gestionesocieta.exception.SocietaConDipendentiAssociatiException;
 import it.prova.gestionesocieta.model.Societa;
 import it.prova.gestionesocieta.repository.SocietaRepository;
 
@@ -45,6 +47,12 @@ public class SocietaServiceImpl implements SocietaService{
 
 	@Transactional
 	public void rimuovi(Societa societaInstance) {
+		
+		TypedQuery<Societa> query=entityManager.createQuery("select s from Societa s join fetch s.dipendenti d where s.id=:id", Societa.class).setParameter("id", societaInstance.getId());
+		if(!query.getResultList().isEmpty()) {
+			throw new SocietaConDipendentiAssociatiException("Societa con almeno un dipendente");
+		}
+
 		societaRepository.delete(societaInstance);
 		
 	}
@@ -62,6 +70,11 @@ public class SocietaServiceImpl implements SocietaService{
 		
 		return entityManager.createQuery(query,Societa.class).getResultList();	
 	
+	}
+
+	@Override
+	public List<Societa> cercaAllDistinctByDipendenti_reditoAnnuoLordoGreaterThan(int value) {
+		return societaRepository.findAllDistinctByDipendenti_reditoAnnuoLordoGreaterThan(value);
 	}
 
 }
